Revert "Merge branch 'development' into feat/CS-42076"
This reverts commit 3978ad3a7e7d69d1156ec7acbdd10a259d4aad10, reversing
changes made to 879896e52be609098dcb737e058cfadd09fbc6d8.
