+++ .gitignore
@@ -16,4 +16,3 @@ contentstack-cli-logs
 packages/**/package-lock.json
 .dccache
 yarn.lock
-contents-*
