+++ packages/contentstack-audit/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@contentstack/cli-audit",
-  "version": "1.2.0",
+  "version": "1.0.0",
   "description": "Contentstack audit plugin",
   "author": "Contentstack CLI",
   "homepage": "https://github.com/contentstack/cli",
@@ -20,21 +20,17 @@
   "dependencies": {
     "@contentstack/cli-command": "~1.2.11",
     "@contentstack/cli-utilities": "~1.5.1",
-    "@oclif/plugin-help": "^5",
+    "@oclif/plugin-help": "^5.2.20",
     "@oclif/plugin-plugins": "^3.8.4",
     "chalk": "^4.1.2",
     "fast-csv": "^4.3.6",
-    "fs-extra": "^11.1.1",
     "lodash": "^4.17.21",
-    "uuid": "^9.0.1",
-    "winston": "^3.10.0"
+    "winston": "^3.9.0"
   },
   "devDependencies": {
     "@oclif/test": "^2.0.3",
     "@types/chai": "^4.3.5",
-    "@types/fs-extra": "^11.0.2",
     "@types/node": "^20.3.1",
-    "@types/uuid": "^9.0.4",
     "chai": "^4.3.7",
     "eslint": "^8.43.0",
     "eslint-config-oclif": "^4.0.0",
@@ -61,6 +57,11 @@
     "additionalVersionFlags": [
       "-v"
     ],
+    "topics": {
+      "cm:stacks:audit": {
+        "description": "Audit and find possible refrence errors in the exported data"
+      }
+    },
     "repositoryPrefix": "<%- repo %>/blob/main/packages/contentstack-audit/<%- commandPath %>"
   },
   "scripts": {
