+++ packages/contentstack-audit/src/modules/entries.ts
@@ -1,28 +1,21 @@
 import find from 'lodash/find';
-import { ux } from '@oclif/core';
 import values from 'lodash/values';
 import isEmpty from 'lodash/isEmpty';
 import { join, resolve } from 'path';
+import { existsSync, readFileSync } from 'fs';
 import { FsUtility } from '@contentstack/cli-utilities';
-import { existsSync, readFileSync, writeFileSync } from 'fs';
 
-import auditConfig from '../config';
-import ContentType from './content-types';
-import { $t, auditMsg, commonMsg } from '../messages';
 import {
   LogFn,
   Locale,
   ConfigType,
   EntryStruct,
-  EntryFieldType,
   ModularBlockType,
   ContentTypeStruct,
   CtConstructorParam,
   GroupFieldDataType,
   GlobalFieldDataType,
   JsonRTEFieldDataType,
-  ContentTypeSchemaType,
-  EntryModularBlockType,
   ModularBlocksDataType,
   ModuleConstructorParam,
   ReferenceFieldDataType,
@@ -33,10 +26,11 @@ import {
   EntryModularBlocksDataType,
   EntryReferenceFieldDataType,
 } from '../types';
+import auditConfig from '../config';
+import { $t, auditMsg } from '../messages';
 
 export default class Entries {
   public log: LogFn;
-  private fix: boolean;
   public fileName: string;
   public locales!: Locale[];
   public config: ConfigType;
@@ -45,20 +39,19 @@ export default class Entries {
   public currentTitle!: string;
   public gfSchema: ContentTypeStruct[];
   public ctSchema: ContentTypeStruct[];
-  private entries!: Record<string, EntryStruct>;
   protected missingRefs: Record<string, any> = {};
   public entryMetaData: Record<string, any>[] = [];
-  public moduleName: keyof typeof auditConfig.moduleConfig = 'entries';
+  public moduleName: keyof typeof auditConfig.moduleConfig = 'content-types';
 
-  constructor({ log, fix, config, moduleName, ctSchema, gfSchema }: ModuleConstructorParam & CtConstructorParam) {
+  constructor({ log, config, moduleName, ctSchema, gfSchema }: ModuleConstructorParam & CtConstructorParam) {
     this.log = log;
     this.config = config;
-    this.fix = fix ?? false;
     this.ctSchema = ctSchema;
     this.gfSchema = gfSchema;
-    this.moduleName = moduleName ?? 'entries';
     this.fileName = config.moduleConfig[this.moduleName].fileName;
     this.folderPath = resolve(config.basePath, config.moduleConfig.entries.dirName);
+
+    if (moduleName) this.moduleName = moduleName;
   }
 
   /**
@@ -73,40 +66,21 @@ export default class Entries {
 
     await this.prepareEntryMetaData();
 
-    this.ctSchema = (await new ContentType({
-      fix: true,
-      log: () => {},
-      config: this.config,
-      moduleName: 'content-types',
-      ctSchema: this.ctSchema,
-      gfSchema: this.gfSchema,
-    }).run(true)) as ContentTypeStruct[];
-    this.gfSchema = (await new ContentType({
-      fix: true,
-      log: () => {},
-      config: this.config,
-      moduleName: 'entries',
-      ctSchema: this.ctSchema,
-      gfSchema: this.gfSchema,
-    }).run(true)) as ContentTypeStruct[];
-
     for (const { code } of this.locales) {
       for (const ctSchema of this.ctSchema) {
         const basePath = join(this.folderPath, ctSchema.uid, code);
         const fsUtility = new FsUtility({ basePath, indexFileName: 'index.json' });
         const indexer = fsUtility.indexFileContent;
 
-        for (const fileIndex in indexer) {
+        for (const _ in indexer) {
           const entries = (await fsUtility.readChunkFiles.next()) as Record<string, EntryStruct>;
-          this.entries = entries;
-
-          for (const entryUid in this.entries) {
-            const entry = this.entries[entryUid];
+          for (const entryUid in entries) {
+            let entry = entries[entryUid];
             const { uid, title } = entry;
             this.currentUid = uid;
             this.currentTitle = title;
             this.missingRefs[this.currentUid] = [];
-            this.lookForReference([{ uid, name: title }], ctSchema, this.entries[entryUid]);
+            await this.lookForReference([{ uid, name: title }], ctSchema, entry);
             this.log(
               $t(auditMsg.SCAN_ENTRY_SUCCESS_MSG, {
                 title,
@@ -115,10 +89,6 @@ export default class Entries {
               }),
               'info',
             );
-          }
-
-          if (this.fix) {
-            await this.writeFixContent(`${basePath}/${indexer[fileIndex]}`, this.entries);
           }
         }
       }
@@ -135,22 +105,6 @@ export default class Entries {
   }
 
   /**
-   * The function checks if it can write the fix content to a file and if so, it writes the content as
-   * JSON to the specified file path.
-   */
-  async writeFixContent(filePath: string, schema: Record<string, EntryStruct>) {
-    let canWrite = true;
-
-    if (this.fix && !this.config.flags['copy-dir']) {
-      canWrite = this.config.flags.yes || (await ux.confirm(commonMsg.FIX_CONFIRMATION));
-    }
-
-    if (canWrite) {
-      writeFileSync(filePath, JSON.stringify(schema));
-    }
-  }
-
-  /**
    * The function `lookForReference` iterates over a given schema and validates different field types
    * such as reference, global field, JSON, modular blocks, and group fields.
    * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure of
@@ -162,61 +116,54 @@ export default class Entries {
    * EntryGroupFieldDataType} entry - The `entry` parameter is an object that represents the data of an
    * entry. It can have different types depending on the `schema` parameter.
    */
-  lookForReference(
+  async lookForReference(
     tree: Record<string, unknown>[],
-    field: ContentTypeStruct | GlobalFieldDataType | ModularBlockType | GroupFieldDataType,
-    entry: EntryFieldType,
-  ) {
-    if (this.fix) {
-      entry = this.runFixOnSchema(tree, field.schema as ContentTypeSchemaType[], entry);
-    }
-
-    for (const child of field.schema ?? []) {
-      const { uid } = child;
-
-      if (!entry?.[uid]) return;
-
-      switch (child.data_type) {
+    { schema }: ContentTypeStruct | GlobalFieldDataType | ModularBlockType | GroupFieldDataType,
+    entry: EntryStruct | EntryGlobalFieldDataType | EntryModularBlocksDataType | EntryGroupFieldDataType,
+  ): Promise<void> {
+    for (const field of schema ?? []) {
+      const { uid } = field;
+      switch (field.data_type) {
         case 'reference':
           this.missingRefs[this.currentUid].push(
             ...this.validateReferenceField(
-              [...tree, { uid: child.uid, name: child.display_name, field: uid }],
-              child as ReferenceFieldDataType,
+              [...tree, { uid: field.uid, name: field.display_name, field: uid }],
+              field as ReferenceFieldDataType,
               entry[uid] as EntryReferenceFieldDataType[],
             ),
           );
           break;
         case 'global_field':
-          this.validateGlobalField(
-            [...tree, { uid: child.uid, name: child.display_name, field: uid }],
-            child as GlobalFieldDataType,
+          await this.validateGlobalField(
+            [...tree, { uid: field.uid, name: field.display_name, field: uid }],
+            field as GlobalFieldDataType,
             entry[uid] as EntryGlobalFieldDataType,
           );
           break;
         case 'json':
-          if (child.field_metadata.extension) {
+          if (field.field_metadata.extension) {
             // NOTE Custom field type
-          } else if (child.field_metadata.allow_json_rte) {
+          } else if (field.field_metadata.allow_json_rte) {
             // NOTE JSON RTE field type
-            this.validateJsonRTEFields(
-              [...tree, { uid: child.uid, name: child.display_name, field: uid }],
-              child as JsonRTEFieldDataType,
+            await this.validateJsonRTEFields(
+              [...tree, { uid: field.uid, name: field.display_name, field: uid }],
+              field as JsonRTEFieldDataType,
               entry[uid] as EntryJsonRTEFieldDataType,
             );
           }
           break;
         case 'blocks':
-          this.validateModularBlocksField(
-            [...tree, { uid: child.uid, name: child.display_name, field: uid }],
-            child as ModularBlocksDataType,
+          await this.validateModularBlocksField(
+            [...tree, { uid: field.uid, name: field.display_name, field: uid }],
+            field as ModularBlocksDataType,
             entry[uid] as EntryModularBlocksDataType[],
           );
           break;
         case 'group':
-          this.validateGroupField(
-            [...tree, { uid: field.uid, name: child.display_name, field: uid }],
-            child as GroupFieldDataType,
-            entry[uid] as EntryGroupFieldDataType[],
+          await this.validateGroupField(
+            [...tree, { uid: field.uid, name: field.display_name, field: uid }],
+            field as GroupFieldDataType,
+            entry[uid] as EntryGroupFieldDataType,
           );
           break;
       }
@@ -256,13 +203,13 @@ export default class Entries {
    * @param {EntryGlobalFieldDataType} field - The `field` parameter is of type
    * `EntryGlobalFieldDataType`. It represents a single global field entry.
    */
-  validateGlobalField(
+  async validateGlobalField(
     tree: Record<string, unknown>[],
     fieldStructure: GlobalFieldDataType,
     field: EntryGlobalFieldDataType,
-  ) {
+  ): Promise<void> {
     // NOTE Any GlobalField related logic can be added here
-    this.lookForReference(tree, fieldStructure, field);
+    await this.lookForReference(tree, fieldStructure, field);
   }
 
   /**
@@ -277,23 +224,38 @@ export default class Entries {
    * `EntryJsonRTEFieldDataType`, which represents a JSON RTE field in an entry. It contains properties
    * such as `uid`, `attrs`, and `children`.
    */
-  validateJsonRTEFields(
+  async validateJsonRTEFields(
     tree: Record<string, unknown>[],
     fieldStructure: JsonRTEFieldDataType,
     field: EntryJsonRTEFieldDataType,
   ) {
-    // const missingRefIndex = []
     // NOTE Other possible reference logic will be added related to JSON RTE (Ex missing assets, extensions etc.,)
-    for (const index in field?.children ?? []) {
-      const child = field.children[index];
-      const { children } = child;
+    for (const child of field?.children ?? []) {
+      const { uid: childrenUid, attrs, children } = child;
+      const { 'entry-uid': entryUid, 'content-type-uid': contentTypeUid } = attrs || {};
 
-      if (!this.fix) {
-        this.jsonRefCheck(tree, fieldStructure, child);
+      if (entryUid) {
+        const refExist = find(this.entryMetaData, { uid: entryUid });
+
+        if (!refExist) {
+          tree.push({ field: 'children' }, { field: childrenUid, uid: fieldStructure.uid });
+          this.missingRefs[this.currentUid].push({
+            tree,
+            uid: this.currentUid,
+            name: this.currentTitle,
+            data_type: fieldStructure.data_type,
+            display_name: fieldStructure.display_name,
+            treeStr: tree
+              .map(({ name }) => name)
+              .filter((val) => val)
+              .join(' ➜ '),
+            missingRefs: [{ uid: entryUid, 'content-type-uid': contentTypeUid }],
+          });
+        }
       }
 
       if (!isEmpty(children)) {
-        this.validateJsonRTEFields(tree, fieldStructure, field.children[index]);
+        await this.validateJsonRTEFields(tree, fieldStructure, child);
       }
     }
   }
@@ -311,24 +273,21 @@ export default class Entries {
    * @param {EntryModularBlocksDataType[]} field - The `field` parameter is an array of objects of type
    * `EntryModularBlocksDataType`.
    */
-  validateModularBlocksField(
+  async validateModularBlocksField(
     tree: Record<string, unknown>[],
     fieldStructure: ModularBlocksDataType,
     field: EntryModularBlocksDataType[],
-  ) {
-    if (!this.fix) {
-      for (const index in field) {
-        this.modularBlockRefCheck(tree, fieldStructure.blocks, field[index], +index);
-      }
-    }
+  ): Promise<void> {
+    const { blocks } = fieldStructure;
 
-    for (const block of fieldStructure.blocks) {
-      const { uid, title } = block;
+    // NOTE Traverse each and every module and look for reference
+    for (let index = 0; index < blocks.length; index++) {
+      const ctBlock = blocks[index];
+      const entryBlock = field[index];
+      const { uid } = ctBlock;
 
-      for (const eBlock of field) {
-        if (eBlock[uid]) {
-          this.lookForReference([...tree, { uid, name: title }], block, eBlock[uid] as EntryModularBlocksDataType);
-        }
+      if (entryBlock?.[uid]) {
+        await this.lookForReference([...tree, { field: uid }], ctBlock, entryBlock[uid] as EntryModularBlocksDataType);
       }
     }
   }
@@ -342,23 +301,13 @@ export default class Entries {
    * @param {EntryGroupFieldDataType} field - The `field` parameter is of type
    * `EntryGroupFieldDataType` and represents a single group field entry.
    */
-  validateGroupField(
+  async validateGroupField(
     tree: Record<string, unknown>[],
     fieldStructure: GroupFieldDataType,
-    field: EntryGroupFieldDataType | EntryGroupFieldDataType[],
-  ) {
+    field: EntryGroupFieldDataType,
+  ): Promise<void> {
     // NOTE Any Group Field related logic can be added here (Ex data serialization or picking any metadata for report etc.,)
-    if (Array.isArray(field)) {
-      field.forEach((eGroup) => {
-        this.lookForReference(
-          [...tree, { uid: fieldStructure.uid, display_name: fieldStructure.display_name }],
-          fieldStructure,
-          eGroup,
-        );
-      });
-    } else {
-      this.lookForReference(tree, fieldStructure, field);
-    }
+    await this.lookForReference(tree, fieldStructure, field);
   }
 
   /**
@@ -381,13 +330,10 @@ export default class Entries {
     fieldStructure: ReferenceFieldDataType,
     field: EntryReferenceFieldDataType[],
   ): EntryRefErrorReturnType[] {
-    if (this.fix) return [];
+    const missingRefs = [];
+    const { data_type, display_name } = fieldStructure;
 
-    const missingRefs: Record<string, any>[] = [];
-    const { uid: data_type, display_name } = fieldStructure;
-
-    for (const index in field ?? []) {
-      const reference = field[index];
+    for (const reference of field ?? []) {
       const { uid } = reference;
       // NOTE Can skip specific references keys (Ex, system defined keys can be skipped)
       // if (this.config.skipRefs.includes(reference)) continue;
@@ -395,7 +341,7 @@ export default class Entries {
       const refExist = find(this.entryMetaData, { uid });
 
       if (!refExist) {
-        missingRefs.push(reference);
+        missingRefs.push(reference as Record<string, unknown>);
       }
     }
 
@@ -415,355 +361,6 @@ export default class Entries {
           },
         ]
       : [];
-  }
-
-  /**
-   * The function `runFixOnSchema` takes in a tree, schema, and entry, and applies fixes to the entry
-   * based on the schema.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure of
-   * the schema. Each object has the following properties:
-   * @param {ContentTypeSchemaType[]} schema - The `schema` parameter is an array of objects
-   * representing the content type schema. Each object in the array contains information about a
-   * specific field in the schema, such as its unique identifier (`uid`) and data type (`data_type`).
-   * @param {EntryFieldType} entry - The `entry` parameter is of type `EntryFieldType`, which
-   * represents the data of an entry. It is an object that contains fields as key-value pairs, where
-   * the key is the field UID (unique identifier) and the value is the field data.
-   * @returns the updated `entry` object after applying fixes to the fields based on the provided
-   * `schema`.
-   */
-  runFixOnSchema(tree: Record<string, unknown>[], schema: ContentTypeSchemaType[], entry: EntryFieldType) {
-    // NOTE Global field Fix
-    schema.forEach((field) => {
-      const { uid, data_type } = field;
-
-      switch (data_type) {
-        case 'global_field':
-          entry[uid] = this.fixGlobalFieldReferences(
-            [...tree, { uid: field.uid, name: field.display_name, data_type: field.data_type }],
-            field as GlobalFieldDataType,
-            entry[uid] as EntryGlobalFieldDataType,
-          ) as EntryGlobalFieldDataType;
-          break;
-        case 'json':
-        case 'reference':
-          if (data_type === 'json') {
-            if (field.field_metadata.extension) {
-              // NOTE Custom field type
-              return field;
-            } else if (field.field_metadata.allow_json_rte) {
-              return this.fixJsonRteMissingReferences(
-                [...tree, { uid: field.uid, name: field.display_name, data_type: field.data_type }],
-                field as JsonRTEFieldDataType,
-                entry[uid] as EntryJsonRTEFieldDataType,
-              );
-            }
-          }
-
-          // NOTE Reference field
-          entry[uid] = this.fixMissingReferences(
-            [...tree, { uid: field.uid, name: field.display_name, data_type: field.data_type }],
-            field as ReferenceFieldDataType,
-            entry[uid] as EntryReferenceFieldDataType[],
-          ) as EntryReferenceFieldDataType[];
-          break;
-        case 'blocks':
-          entry[uid] = this.fixModularBlocksReferences(
-            [...tree, { uid: field.uid, name: field.display_name, data_type: field.data_type }],
-            (field as ModularBlocksDataType).blocks,
-            entry[uid] as EntryModularBlocksDataType[],
-          );
-          break;
-        case 'group':
-          entry[uid] = this.fixGroupField(
-            [...tree, { uid: field.uid, name: field.display_name, data_type: field.data_type }],
-            field as GroupFieldDataType,
-            entry[uid] as EntryGroupFieldDataType[],
-          ) as EntryGroupFieldDataType;
-          break;
-      }
-    });
-
-    return entry;
-  }
-
-  /**
-   * The function `fixGlobalFieldReferences` adds a new entry to a tree data structure and runs a fix
-   * on the schema.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure.
-   * @param {GlobalFieldDataType} field - The `field` parameter is of type `GlobalFieldDataType` and
-   * represents a global field object. It contains properties such as `uid` and `display_name`.
-   * @param {EntryGlobalFieldDataType} entry - The `entry` parameter is of type
-   * `EntryGlobalFieldDataType` and represents the global field entry that needs to be fixed.
-   * @returns the result of calling the `runFixOnSchema` method with the updated `tree` array,
-   * `field.schema`, and `entry` as arguments.
-   */
-  fixGlobalFieldReferences(
-    tree: Record<string, unknown>[],
-    field: GlobalFieldDataType,
-    entry: EntryGlobalFieldDataType,
-  ) {
-    return this.runFixOnSchema([...tree, { uid: field.uid, display_name: field.display_name }], field.schema, entry);
-  }
-
-  /**
-   * The function `fixModularBlocksReferences` takes in a tree, a list of blocks, and an entry, and
-   * performs various operations to fix references within the entry.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure of
-   * the modular blocks.
-   * @param {ModularBlockType[]} blocks - An array of objects representing modular blocks. Each object
-   * has properties like `uid` (unique identifier) and `title` (display name).
-   * @param {EntryModularBlocksDataType[]} entry - An array of objects representing the modular blocks
-   * data in an entry. Each object in the array represents a modular block and contains its unique
-   * identifier (uid) and other properties.
-   * @returns the updated `entry` array after performing some modifications.
-   */
-  fixModularBlocksReferences(
-    tree: Record<string, unknown>[],
-    blocks: ModularBlockType[],
-    entry: EntryModularBlocksDataType[],
-  ) {
-    entry = entry
-      .map((block, index) => this.modularBlockRefCheck(tree, blocks, block, index))
-      .filter((val) => !isEmpty(val));
-
-    blocks.forEach((block) => {
-      entry = entry
-        .map((eBlock) => {
-          if (!isEmpty(block.schema)) {
-            if (eBlock[block.uid]) {
-              eBlock[block.uid] = this.runFixOnSchema(
-                [...tree, { uid: block.uid, display_name: block.title }],
-                block.schema as ContentTypeSchemaType[],
-                eBlock[block.uid] as EntryFieldType,
-              ) as EntryModularBlockType;
-            }
-          }
-
-          return eBlock;
-        })
-        .filter((val) => !isEmpty(val)) as EntryModularBlocksDataType[];
-    });
-
-    return entry;
-  }
-
-  /**
-   * The function `fixGroupField` takes in a tree, a field, and an entry, and if the field has a
-   * schema, it runs a fix on the schema and returns the updated entry, otherwise it returns the
-   * original entry.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure.
-   * @param {GroupFieldDataType} field - The `field` parameter is of type `GroupFieldDataType` and
-   * represents a group field object. It contains properties such as `uid` (unique identifier) and
-   * `display_name` (name of the field).
-   * @param {EntryGroupFieldDataType} entry - The `entry` parameter is of type
-   * `EntryGroupFieldDataType`.
-   * @returns If the `field.schema` is not empty, the function will return the result of calling
-   * `this.runFixOnSchema` with the updated `tree`, `field.schema`, and `entry` as arguments.
-   * Otherwise, it will return the `entry` as is.
-   */
-  fixGroupField(
-    tree: Record<string, unknown>[],
-    field: GroupFieldDataType,
-    entry: EntryGroupFieldDataType | EntryGroupFieldDataType[],
-  ) {
-    if (!isEmpty(field.schema)) {
-      if (Array.isArray(entry)) {
-        entry = entry.map((eGroup) => {
-          return this.runFixOnSchema(
-            [...tree, { uid: field.uid, display_name: field.display_name }],
-            field.schema as ContentTypeSchemaType[],
-            eGroup,
-          );
-        }) as EntryGroupFieldDataType[];
-      } else {
-        entry = this.runFixOnSchema(
-          [...tree, { uid: field.uid, display_name: field.display_name }],
-          field.schema as ContentTypeSchemaType[],
-          entry,
-        ) as EntryGroupFieldDataType;
-      }
-    }
-
-    return entry;
-  }
-
-  /**
-   * The function fixes missing references in a JSON tree structure.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing a tree structure. Each
-   * object in the array has a string key and an unknown value.
-   * @param {ReferenceFieldDataType | JsonRTEFieldDataType} field - The `field` parameter can be of
-   * type `ReferenceFieldDataType` or `JsonRTEFieldDataType`.
-   * @param {EntryJsonRTEFieldDataType} entry - The `entry` parameter is of type
-   * `EntryJsonRTEFieldDataType`, which represents an entry in a JSON Rich Text Editor (JsonRTE) field.
-   * @returns the updated `entry` object with fixed missing references in the `children` property.
-   */
-  fixJsonRteMissingReferences(
-    tree: Record<string, unknown>[],
-    field: ReferenceFieldDataType | JsonRTEFieldDataType,
-    entry: EntryJsonRTEFieldDataType | EntryJsonRTEFieldDataType[],
-  ) {
-    if (Array.isArray(entry)) {
-      entry = entry.map((child: any, index) => {
-        return this.fixJsonRteMissingReferences(
-          [...tree, { index, type: (child as any)?.type, uid: child?.uid }],
-          field,
-          child,
-        );
-      }) as EntryJsonRTEFieldDataType[];
-    } else {
-      entry.children = entry.children
-        .map((child) => {
-          const refExist = this.jsonRefCheck(tree, field, child);
-
-          if (!refExist) return null;
-
-          if (isEmpty(child.children)) {
-            child = this.fixJsonRteMissingReferences(tree, field, child) as EntryJsonRTEFieldDataType;
-          }
-
-          return child;
-        })
-        .filter((val) => val) as EntryJsonRTEFieldDataType[];
-    }
-
-    return entry;
-  }
-
-  /**
-   * The `fixMissingReferences` function checks for missing references in an entry and adds them to a
-   * list if they are not found.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing a tree structure. Each
-   * object in the array should have a "name" property and an optional "index" property.
-   * @param {ReferenceFieldDataType | JsonRTEFieldDataType} field - The `field` parameter is of type
-   * `ReferenceFieldDataType` or `JsonRTEFieldDataType`.
-   * @param {EntryReferenceFieldDataType[]} entry - The `entry` parameter is an array of objects that
-   * represent references to other entries. Each object in the array has the following properties:
-   * @returns the `entry` variable.
-   */
-  fixMissingReferences(
-    tree: Record<string, unknown>[],
-    field: ReferenceFieldDataType | JsonRTEFieldDataType,
-    entry: EntryReferenceFieldDataType[],
-  ) {
-    const missingRefs: Record<string, any>[] = [];
-    entry = entry
-      .map((reference) => {
-        const { uid } = reference;
-        const refExist = find(this.entryMetaData, { uid });
-
-        if (!refExist) {
-          missingRefs.push(reference);
-          return null;
-        }
-
-        return reference;
-      })
-      .filter((val) => val) as EntryReferenceFieldDataType[];
-
-    if (!isEmpty(missingRefs)) {
-      this.missingRefs[this.currentUid].push({
-        tree,
-        fixStatus: 'Fixed',
-        uid: this.currentUid,
-        name: this.currentTitle,
-        data_type: field.data_type,
-        display_name: field.display_name,
-        treeStr: tree
-          .map(({ name, index }) => (index || index === 0 ? `[${index}].${name}` : name))
-          .filter((val) => val)
-          .join(' ➜ '),
-        missingRefs,
-      });
-    }
-
-    return entry;
-  }
-
-  /**
-   * The function `modularBlockRefCheck` checks for invalid keys in an entry block and returns the
-   * updated entry block.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure of
-   * the blocks.
-   * @param {ModularBlockType[]} blocks - The `blocks` parameter is an array of `ModularBlockType`
-   * objects.
-   * @param {EntryModularBlocksDataType} entryBlock - The `entryBlock` parameter is an object that
-   * represents a modular block entry. It contains key-value pairs where the keys are the UIDs of the
-   * modular blocks and the values are the data associated with each modular block.
-   * @param {Number} index - The `index` parameter is a number that represents the index of the current
-   * block in the `tree` array.
-   * @returns the `entryBlock` object.
-   */
-  modularBlockRefCheck(
-    tree: Record<string, unknown>[],
-    blocks: ModularBlockType[],
-    entryBlock: EntryModularBlocksDataType,
-    index: Number,
-  ) {
-    const validBlockUid = blocks.map((block) => block.uid);
-    const invalidKeys = Object.keys(entryBlock).filter((key) => !validBlockUid.includes(key));
-
-    invalidKeys.forEach((key) => {
-      if (this.fix) {
-        delete entryBlock[key];
-      }
-
-      this.missingRefs[this.currentUid].push({
-        uid: this.currentUid,
-        name: this.currentTitle,
-        data_type: key,
-        display_name: key,
-        fixStatus: this.fix ? 'Fixed' : undefined,
-        tree: [...tree, { index, uid: key, name: key }],
-        treeStr: [...tree, { index, uid: key, name: key }]
-          .map(({ name, index }) => (index || index === 0 ? `[${index}].${name}` : name))
-          .filter((val) => val)
-          .join(' ➜ '),
-        missingRefs: [key],
-      });
-    });
-
-    return entryBlock;
-  }
-
-  /**
-   * The `jsonRefCheck` function checks if a reference exists in a JSON tree and adds missing
-   * references to a list if they are not found.
-   * @param {Record<string, unknown>[]} tree - An array of objects representing the tree structure.
-   * @param {JsonRTEFieldDataType} schema - The `schema` parameter is of type `JsonRTEFieldDataType`
-   * and represents the schema of a JSON field. It contains properties such as `uid`, `data_type`, and
-   * `display_name`.
-   * @param {EntryJsonRTEFieldDataType} child - The `child` parameter is an object that represents a
-   * child entry in a JSON tree. It has the following properties:
-   * @returns The function `jsonRefCheck` returns either `null` or `true`.
-   */
-  jsonRefCheck(tree: Record<string, unknown>[], schema: JsonRTEFieldDataType, child: EntryJsonRTEFieldDataType) {
-    const { uid: childrenUid } = child;
-    const { 'entry-uid': entryUid, 'content-type-uid': contentTypeUid } = child.attrs || {};
-
-    if (entryUid) {
-      const refExist = find(this.entryMetaData, { uid: entryUid });
-
-      if (!refExist) {
-        tree.push({ field: 'children' }, { field: childrenUid, uid: schema.uid });
-        this.missingRefs[this.currentUid].push({
-          tree,
-          uid: this.currentUid,
-          name: this.currentTitle,
-          data_type: schema.data_type,
-          display_name: schema.display_name,
-          fixStatus: this.fix ? 'Fixed' : undefined,
-          treeStr: tree
-            .map(({ name }) => name)
-            .filter((val) => val)
-            .join(' ➜ '),
-          missingRefs: [{ uid: entryUid, 'content-type-uid': contentTypeUid }],
-        });
-
-        return null;
-      }
-    }
-
-    return true;
   }
 
   /**
