+++ packages/contentstack-audit/src/types/content-types.ts
@@ -1,24 +1,22 @@
 import config from '../config';
 import { ConfigType, LogFn } from './utils';
 
-type ContentTypeSchemaType =
-  | ReferenceFieldDataType
-  | GlobalFieldDataType
-  | CustomFieldDataType
-  | JsonRTEFieldDataType
-  | GroupFieldDataType
-  | ModularBlocksDataType;
-
 type ContentTypeStruct = {
   uid: string;
   title: string;
   description: string;
-  schema?: ContentTypeSchemaType[];
+  schema?: (
+    | ReferenceFieldDataType
+    | GlobalFieldDataType
+    | CustomFieldDataType
+    | JsonRTEFieldDataType
+    | GroupFieldDataType
+    | ModularBlocksDataType
+  )[];
 };
 
 type ModuleConstructorParam = {
   log: LogFn;
-  fix?: boolean
   config: ConfigType;
   moduleName?: keyof typeof config.moduleConfig;
 };
@@ -44,7 +42,6 @@ type RefErrorReturnType = {
   ct_uid: string;
   treeStr: string;
   data_type: string;
-  fixStatus?: string
   missingRefs: string[];
   display_name: string;
   tree: Record<string, unknown>[];
@@ -57,15 +54,12 @@ type ReferenceFieldDataType = CommonDataTypeStruct & {
 
 // NOTE Type 2
 type GlobalFieldDataType = CommonDataTypeStruct & {
-  reference_to?: string;
+  reference_to: string;
   schema: GlobalFieldSchemaTypes[];
 };
 
 // NOTE Type 3
-type CustomFieldDataType = CommonDataTypeStruct & {
-  reference_to: string[];
-  extension_uid: string;
-};
+type CustomFieldDataType = CommonDataTypeStruct & {};
 
 // NOTE Type 4
 type JsonRTEFieldDataType = CommonDataTypeStruct & {
@@ -81,7 +75,6 @@ type GroupFieldDataType = CommonDataTypeStruct & {
 type ModularBlockType = {
   uid: string;
   title: string;
-  reference_to?: string;
   schema: (
     | JsonRTEFieldDataType
     | ModularBlocksDataType
@@ -107,7 +100,7 @@ enum OutputColumn {
   'Field name' = 'display_name',
   'Field type' = 'data_type',
   'Missing references' = 'missingRefs',
-  Path = 'treeStr'
+  Path = 'treeStr',
 }
 
 export {
@@ -124,5 +117,4 @@ export {
   ModularBlocksSchemaTypes,
   ModularBlockType,
   OutputColumn,
-  ContentTypeSchemaType,
 };
