+++ packages/contentstack-audit/src/types/entries.ts
@@ -41,10 +41,8 @@ type EntryJsonRTEFieldDataType = {
 };
 
 // NOTE Type 5
-type GroupFieldType = EntryReferenceFieldDataType[] | EntryGlobalFieldDataType | EntryJsonRTEFieldDataType;
-
 type EntryGroupFieldDataType = {
-  [key: string]: GroupFieldType;
+  [key: string]: EntryReferenceFieldDataType[] | EntryGlobalFieldDataType | EntryJsonRTEFieldDataType;
 };
 
 // NOTE Type 6
@@ -65,18 +63,14 @@ type EntryRefErrorReturnType = {
   uid: string;
   treeStr: string;
   data_type: string;
-  fixStatus?: string;
+  missingRefs: string[] | Record<string, unknown>[];
   display_name: string;
   tree: Record<string, unknown>[];
-  missingRefs: string[] | Record<string, unknown>[];
 };
-
-type EntryFieldType = EntryStruct | EntryGlobalFieldDataType | EntryModularBlocksDataType | EntryGroupFieldDataType;
 
 export {
   Locale,
   EntryStruct,
-  EntryFieldType,
   EntryGlobalFieldDataType,
   EntryCustomFieldDataType,
   EntryJsonRTEFieldDataType,
@@ -84,6 +78,4 @@ export {
   EntryModularBlocksDataType,
   EntryReferenceFieldDataType,
   EntryRefErrorReturnType,
-  GroupFieldType,
-  EntryModularBlockType,
 };
