+++ packages/contentstack-bootstrap/README.md
@@ -15,7 +15,7 @@ $ npm install -g @contentstack/cli-cm-bootstrap
 $ csdx COMMAND
 running command...
 $ csdx (--version)
-@contentstack/cli-cm-bootstrap/1.5.1 darwin-arm64 node-v20.8.0
+@contentstack/cli-cm-bootstrap/1.5.1 darwin-x64 node-v20.8.0
 $ csdx --help [COMMAND]
 USAGE
   $ csdx COMMAND
