+++ packages/contentstack-bulk-publish/README.md
@@ -18,7 +18,7 @@ $ npm install -g @contentstack/cli-cm-bulk-publish
 $ csdx COMMAND
 running command...
 $ csdx (--version)
-@contentstack/cli-cm-bulk-publish/1.3.12 darwin-arm64 node-v20.8.0
+@contentstack/cli-cm-bulk-publish/1.3.12 darwin-x64 node-v20.8.0
 $ csdx --help [COMMAND]
 USAGE
   $ csdx COMMAND
