+++ packages/contentstack-config/README.md
@@ -18,7 +18,7 @@ $ npm install -g @contentstack/cli-config
 $ csdx COMMAND
 running command...
 $ csdx (--version)
-@contentstack/cli-config/1.4.12 darwin-arm64 node-v20.8.0
+@contentstack/cli-config/1.4.12 darwin-x64 node-v20.8.0
 $ csdx --help [COMMAND]
 USAGE
   $ csdx COMMAND
