+++ packages/contentstack-migrate-rte/README.md
@@ -16,7 +16,7 @@ $ npm install -g @contentstack/cli-cm-migrate-rte
 $ csdx COMMAND
 running command...
 $ csdx (--version)
-@contentstack/cli-cm-migrate-rte/1.4.12 darwin-arm64 node-v20.8.0
+@contentstack/cli-cm-migrate-rte/1.4.12 darwin-x64 node-v20.8.0
 $ csdx --help [COMMAND]
 USAGE
   $ csdx COMMAND
