+++ packages/contentstack/package.json
@@ -22,7 +22,7 @@
     "prepack": "pnpm compile && oclif manifest && oclif readme"
   },
   "dependencies": {
-    "@contentstack/cli-audit": "~1.2.0",
+    "@contentstack/cli-audit": "~1.0.0",
     "@contentstack/cli-auth": "~1.3.14",
     "@contentstack/cli-cm-bootstrap": "~1.5.1",
     "@contentstack/cli-cm-branches": "~1.0.14",
